requirements.txt
Django==1.5.5
Markdown==2.4
Pillow==2.3.0
South==0.8.4
argparse==1.2.1
cmsplugin-filer==0.9.5
cmsplugin-image-gallery==0.6
cmsplugin-markdown==0.1.1
django-admin-jqueryui==1.8.22
django-classy-tags==0.5
django-cms==2.4.3
django-filer==0.9.5
django-inline-ordering==1.0.2
django-jquery==1.9.1
django-mptt==0.5.2
django-polymorphic==0.5.3
django-sekizai==0.7
djangocms-admin-style==0.2.0
djangocms-text-ckeditor==1.0.11
easy-thumbnails==1.5
html5lib==0.999
psycopg2==2.5.2
six==1.5.2
textile==2.1.5
wsgiref==0.1.2
